INS-1979: remove genesis initializer from components, explicitly use genesis.BaseRecord for base record creation.
