+++ ledger/genesis/genesis.go
@@ -20,133 +20,128 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/insolar/insolar/bootstrap/genesis"
-	"github.com/insolar/insolar/component"
-	"github.com/insolar/insolar/configuration"
+	"github.com/pkg/errors"
+
 	"github.com/insolar/insolar/insolar"
+	"github.com/insolar/insolar/insolar/record"
 	"github.com/insolar/insolar/instrumentation/inslogger"
-	"github.com/insolar/insolar/internal/ledger/artifact"
-	storagegenesis "github.com/insolar/insolar/ledger/storage/genesis"
-	"github.com/insolar/insolar/log"
+	"github.com/insolar/insolar/internal/ledger/store"
+	"github.com/insolar/insolar/ledger/storage"
+	"github.com/insolar/insolar/ledger/storage/drop"
+	"github.com/insolar/insolar/ledger/storage/object"
+	"github.com/insolar/insolar/ledger/storage/pulse"
 )
 
-type Initializer struct {
-	cfgPath           string
-	genesisConfigPath string
-	genesisKeyOut     string
+// BaseRecord provides methods for genesis base record manipulation.
+type BaseRecord struct {
+	PCS insolar.PlatformCryptographyScheme
+
+	DB             store.DB
+	DropModifier   drop.Modifier
+	PulseAppender  pulse.Appender
+	PulseAccessor  pulse.Accessor
+	RecordModifier object.RecordModifier
+	// TODO: @imarkin 28.03.2019 - remove ObjectStorage after all new storages integration (INS-2013, etc)
+	ObjectStorage storage.ObjectStorage
 }
 
-func NewInitializer(cfgPath string, genesisConfigPath, genesisKeyOut string) *Initializer {
-	return &Initializer{
-		cfgPath:           cfgPath,
-		genesisConfigPath: genesisConfigPath,
-		genesisKeyOut:     genesisKeyOut,
-	}
+// Key is genesis key.
+type Key struct{}
+
+func (Key) ID() []byte {
+	return []byte{0x01}
 }
 
-func (s *Initializer) Run() {
-	cfgHolder := configuration.NewHolder()
-	var err error
-	if len(s.cfgPath) != 0 {
-		err = cfgHolder.LoadFromFile(s.cfgPath)
-	} else {
-		err = cfgHolder.Load()
+func (Key) Scope() store.Scope {
+	return store.ScopeSystem
+}
+
+// CreateIfNeeded creates new base genesis record if needed.
+// Returns reference of genesis record and flag if base record have been created.
+func (gi *BaseRecord) CreateIfNeeded(ctx context.Context) (*insolar.Reference, bool, error) {
+	inslog := inslogger.FromContext(ctx)
+	inslog.Info("start storage bootstrap")
+
+	getGenesisRef := func() (*insolar.Reference, error) {
+		buff, err := gi.DB.Get(Key{})
+		if err != nil {
+			return nil, err
+		}
+		var genesisRef insolar.Reference
+		copy(genesisRef[:], buff)
+		return &genesisRef, nil
 	}
-	if err != nil {
-		// TODO: should be fatal error
-		log.Warn("failed to load configuration from file: ", err.Error())
-	}
 
-	cfg := &cfgHolder.Configuration
-	ctx, inslog := initLogger(context.Background(), cfg.Log)
-	log.SetGlobalLogger(inslog)
-	fmt.Println("Starts with configuration:\n", configuration.ToString(cfgHolder.Configuration))
-
-	geneisConfig, err := genesis.ParseGenesisConfig(s.genesisConfigPath)
-	checkError(ctx, err, "failed to create genesis Generator")
-
-	bc := initBootstrapComponents(ctx, *cfg)
-	certManager := createCertificateManager(
-		ctx,
-		bc.CryptographyService,
-		bc.KeyProcessor,
-	)
-
-	genesisInitializer := storagegenesis.NewGenesisInitializer()
-
-	sc := initStorageComponents(cfg.Ledger)
-
-	fmt.Println("Initializer: Inject")
-	cm := component.Manager{}
-	cm.Inject(
-		bc.PlatformCryptographyScheme,
-		bc.KeyStore,
-		bc.CryptographyService,
-		bc.KeyProcessor,
-		certManager,
-
-		sc.blobDB,
-		sc.dropDB,
-		sc.recordDB,
-		sc.storageDBContext,
-		sc.storeBadgerDB,
-		sc.objectStorage,
-		sc.pulseDB,
-
-		genesisInitializer,
-	)
-
-	err = cm.Init(ctx)
-	checkError(ctx, err, "failed to init components")
-
-	err = cm.Start(ctx)
-	checkError(ctx, err, "failed to start components")
-
-	// run generator if genesisInitializer happened
-	if genesisInitializer.Happened() {
-		artifactManager := &artifact.Scope{
-			PulseNumber: insolar.FirstPulseNumber,
-
-			PlatformCryptographyScheme: bc.PlatformCryptographyScheme,
-			BlobModifier:               sc.blobDB,
-			RecordsModifier:            sc.recordDB,
-
-			ObjectStorage: sc.objectStorage,
+	createGenesisRecord := func() (*insolar.Reference, error) {
+		err := gi.PulseAppender.Append(
+			ctx,
+			insolar.Pulse{
+				PulseNumber: insolar.GenesisPulse.PulseNumber,
+				Entropy:     insolar.GenesisPulse.Entropy,
+			},
+		)
+		if err != nil {
+			return nil, errors.Wrap(err, "fail to set genesis pulse")
+		}
+		// Add initial drop
+		err = gi.DropModifier.Set(ctx, drop.Drop{JetID: insolar.ZeroJetID})
+		if err != nil {
+			return nil, errors.Wrap(err, "fail to set initial drop")
 		}
 
-		genesisGenerator := genesis.NewGenerator(
-			geneisConfig,
-			artifactManager,
-			*genesisInitializer.GenesisRef(),
-			s.genesisKeyOut,
+		lastPulse, err := gi.PulseAccessor.Latest(ctx)
+		if err != nil {
+			return nil, errors.Wrap(err, "fail to get last pulse")
+		}
+		if lastPulse.PulseNumber != insolar.GenesisPulse.PulseNumber {
+			return nil, fmt.Errorf(
+				"last pulse number %v is not equal to genesis special value %v",
+				lastPulse.PulseNumber,
+				insolar.GenesisPulse.PulseNumber,
+			)
+		}
+
+		virtRec := &object.GenesisRecord{}
+		genesisID := object.NewRecordIDFromRecord(gi.PCS, lastPulse.PulseNumber, virtRec)
+		rec := record.MaterialRecord{
+			Record: virtRec,
+			JetID:  insolar.ZeroJetID,
+		}
+		err = gi.RecordModifier.Set(ctx, *genesisID, rec)
+		if err != nil {
+			return nil, errors.Wrap(err, "can't save genesis record into storage")
+		}
+
+		err = gi.ObjectStorage.SetObjectIndex(
+			ctx,
+			insolar.ID(insolar.ZeroJetID),
+			genesisID,
+			&object.Lifeline{
+				LatestState:         genesisID,
+				LatestStateApproved: genesisID,
+			},
 		)
-		err = genesisGenerator.Run(ctx)
-		checkError(ctx, err, "failed to generate genesis")
+		if err != nil {
+			return nil, errors.Wrap(err, "fail to set genesis index")
+		}
+
+		genesisRef := insolar.NewReference(*genesisID, *genesisID)
+		return genesisRef, gi.DB.Set(Key{}, genesisRef[:])
 	}
 
-	err = cm.Stop(ctx)
-	checkError(ctx, err, "failed to stop components")
-}
-
-func initLogger(ctx context.Context, cfg configuration.Log) (context.Context, insolar.Logger) {
-	inslog, err := log.NewLog(cfg)
-	if err != nil {
-		panic(err)
-	}
-
-	if newInslog, err := inslog.WithLevel(cfg.Level); err != nil {
-		inslog.Error(err.Error())
-	} else {
-		inslog = newInslog
-	}
-
-	ctx = inslogger.SetLogger(ctx, inslog)
-	return ctx, inslog
-}
-
-func checkError(ctx context.Context, err error, message string) {
+	var err error
+	genesisRef, err := getGenesisRef()
 	if err == nil {
-		return
+		return genesisRef, false, nil
 	}
-	inslogger.FromContext(ctx).Fatalf("%v: %v", message, err.Error())
+	if err != store.ErrNotFound {
+		return nil, false, errors.Wrap(err, "genesis bootstrap failed")
+	}
+
+	genesisRef, err = createGenesisRecord()
+	if err != nil {
+		return nil, true, err
+	}
+
+	return genesisRef, true, nil
 }
