--- ledger/genesis/init.go
@@ -1,1 +1,0 @@
-package genesis
